+++ src/change.h
@@ -19,9 +19,9 @@ void end_change_chain(void);
 bool undo(void);
 bool redo(unsigned int change_id);
 void free_changes(Change *head);
-void buffer_insert_bytes(const char *buf, long len);
-void buffer_delete_bytes(long len);
-void buffer_erase_bytes(long len);
-void buffer_replace_bytes(long del_count, const char *inserted, long ins_count);
+void buffer_insert_bytes(const char *buf, size_t len);
+void buffer_delete_bytes(size_t len);
+void buffer_erase_bytes(size_t len);
+void buffer_replace_bytes(size_t del_count, const char *inserted, size_t ins_count);
 
 #endif
